Reorganised React components.
