+++ bpaotu/bpaotu/frontend/src/pages/search_page/components/search_errors.tsx
@@ -1,12 +1,10 @@
 import * as React from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
 
 import {
     Alert,
 } from 'reactstrap';
 
-const SearchErrors = ({ errors }) => {
+export default ({ errors }) => {
     if (errors.length == 0) {
         return (<span></span>);
     }
@@ -20,11 +18,3 @@ const SearchErrors = ({ errors }) => {
         </Alert>
     );
 };
-
-function mapStateToProps(state) {
-    return {
-        errors: state.searchPage.results.errors,
-    };
-}
-
-export default connect(mapStateToProps, null)(SearchErrors);
